Guarantee 64 bits for s64 on windows

+++ src/types.h
@@ -4,5 +4,5 @@ typedef short int s16;
 typedef unsigned short int u16;
 typedef unsigned int u32;
 typedef int s32;
-typedef long int s64;
+typedef long long int s64;
 typedef s32 FP;
